+++ src/components/DateInput.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { forbidExtraProps, nonNegativeInteger } from 'airbnb-prop-types';
-import { withStyles, withStylesPropTypes } from 'react-with-styles';
+import { css, withStyles, withStylesPropTypes } from 'react-with-styles';
 import throttle from 'lodash/throttle';
 import isTouchDevice from 'is-touch-device';
 
@@ -190,7 +190,6 @@ class DateInput extends React.PureComponent {
       block,
       styles,
       theme: { reactDates },
-      css,
     } = this.props;
 
     const value = dateString || displayValue || '';