+++ src/components/DayPickerKeyboardShortcuts.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { forbidExtraProps } from 'airbnb-prop-types';
-import { withStyles, withStylesPropTypes } from 'react-with-styles';
+import { css, withStyles, withStylesPropTypes } from 'react-with-styles';
 
 import { DayPickerKeyboardShortcutsPhrases } from '../defaultPhrases';
 import getPhrasePropTypes from '../utils/getPhrasePropTypes';
@@ -23,6 +23,7 @@ const propTypes = forbidExtraProps({
   closeKeyboardShortcutsPanel: PropTypes.func,
   phrases: PropTypes.shape(getPhrasePropTypes(DayPickerKeyboardShortcutsPhrases)),
   renderKeyboardShortcutsButton: PropTypes.func,
+  renderKeyboardShortcutsPanel: PropTypes.func,
 });
 
 const defaultProps = {
@@ -33,6 +34,7 @@ const defaultProps = {
   closeKeyboardShortcutsPanel() {},
   phrases: DayPickerKeyboardShortcutsPhrases,
   renderKeyboardShortcutsButton: undefined,
+  renderKeyboardShortcutsPanel: undefined,
 };
 
 function getKeyboardShortcuts(phrases) {
@@ -162,12 +164,12 @@ class DayPickerKeyboardShortcuts extends React.PureComponent {
     const {
       block,
       buttonLocation,
-      css,
       showKeyboardShortcutsPanel,
       closeKeyboardShortcutsPanel,
       styles,
       phrases,
       renderKeyboardShortcutsButton,
+      renderKeyboardShortcutsPanel,
     } = this.props;
 
     const toggleButtonText = showKeyboardShortcutsPanel
@@ -217,49 +219,59 @@ class DayPickerKeyboardShortcuts extends React.PureComponent {
           </button>
         )}
         {showKeyboardShortcutsPanel && (
-          <div
-            {...css(styles.DayPickerKeyboardShortcuts_panel)}
-            role="dialog"
-            aria-labelledby="DayPickerKeyboardShortcuts_title"
-            aria-describedby="DayPickerKeyboardShortcuts_description"
-          >
+          renderKeyboardShortcutsPanel ? (
+            renderKeyboardShortcutsPanel({
+              closeButtonAriaLabel: phrases.hideKeyboardShortcutsPanel,
+              keyboardShortcuts: this.keyboardShortcuts,
+              onCloseButtonClick: closeKeyboardShortcutsPanel,
+              onKeyDown: this.onKeyDown,
+              title: phrases.keyboardShortcuts,
+            })
+          ) : (
             <div
-              {...css(styles.DayPickerKeyboardShortcuts_title)}
-              id="DayPickerKeyboardShortcuts_title"
+              {...css(styles.DayPickerKeyboardShortcuts_panel)}
+              role="dialog"
+              aria-labelledby="DayPickerKeyboardShortcuts_title"
+              aria-describedby="DayPickerKeyboardShortcuts_description"
             >
-              {phrases.keyboardShortcuts}
+              <div
+                {...css(styles.DayPickerKeyboardShortcuts_title)}
+                id="DayPickerKeyboardShortcuts_title"
+              >
+                {phrases.keyboardShortcuts}
+              </div>
+
+              <button
+                ref={this.setHideKeyboardShortcutsButtonRef}
+                {...css(
+                  styles.DayPickerKeyboardShortcuts_buttonReset,
+                  styles.DayPickerKeyboardShortcuts_close,
+                )}
+                type="button"
+                tabIndex="0"
+                aria-label={phrases.hideKeyboardShortcutsPanel}
+                onClick={closeKeyboardShortcutsPanel}
+                onKeyDown={this.onKeyDown}
+              >
+                <CloseButton {...css(styles.DayPickerKeyboardShortcuts_closeSvg)} />
+              </button>
+
+              <ul
+                {...css(styles.DayPickerKeyboardShortcuts_list)}
+                id="DayPickerKeyboardShortcuts_description"
+              >
+                {this.keyboardShortcuts.map(({ unicode, label, action }) => (
+                  <KeyboardShortcutRow
+                    key={label}
+                    unicode={unicode}
+                    label={label}
+                    action={action}
+                    block={block}
+                  />
+                ))}
+              </ul>
             </div>
-
-            <button
-              ref={this.setHideKeyboardShortcutsButtonRef}
-              {...css(
-                styles.DayPickerKeyboardShortcuts_buttonReset,
-                styles.DayPickerKeyboardShortcuts_close,
-              )}
-              type="button"
-              tabIndex="0"
-              aria-label={phrases.hideKeyboardShortcutsPanel}
-              onClick={closeKeyboardShortcutsPanel}
-              onKeyDown={this.onKeyDown}
-            >
-              <CloseButton {...css(styles.DayPickerKeyboardShortcuts_closeSvg)} />
-            </button>
-
-            <ul
-              {...css(styles.DayPickerKeyboardShortcuts_list)}
-              id="DayPickerKeyboardShortcuts_description"
-            >
-              {this.keyboardShortcuts.map(({ unicode, label, action }) => (
-                <KeyboardShortcutRow
-                  key={label}
-                  unicode={unicode}
-                  label={label}
-                  action={action}
-                  block={block}
-                />
-              ))}
-            </ul>
-          </div>
+          )
         )}
       </div>
     );