+++ src/components/DayPickerNavigation.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import { forbidExtraProps } from 'airbnb-prop-types';
-import { withStyles, withStylesPropTypes } from 'react-with-styles';
+import { css, withStyles, withStylesPropTypes } from 'react-with-styles';
 
 import { DayPickerNavigationPhrases } from '../defaultPhrases';
 import getPhrasePropTypes from '../utils/getPhrasePropTypes';
@@ -11,10 +11,13 @@ import LeftArrow from './LeftArrow';
 import RightArrow from './RightArrow';
 import ChevronUp from './ChevronUp';
 import ChevronDown from './ChevronDown';
+import NavPositionShape from '../shapes/NavPositionShape';
 import ScrollableOrientationShape from '../shapes/ScrollableOrientationShape';
 
 import {
   HORIZONTAL_ORIENTATION,
+  NAV_POSITION_BOTTOM,
+  NAV_POSITION_TOP,
   VERTICAL_SCROLLABLE,
 } from '../constants';
 
@@ -22,6 +25,7 @@ const propTypes = forbidExtraProps({
   ...withStylesPropTypes,
   disablePrev: PropTypes.bool,
   disableNext: PropTypes.bool,
+  navPosition: NavPositionShape,
   navPrev: PropTypes.node,
   navNext: PropTypes.node,
   orientation: ScrollableOrientationShape,
@@ -32,12 +36,14 @@ const propTypes = forbidExtraProps({
   // internationalization
   phrases: PropTypes.shape(getPhrasePropTypes(DayPickerNavigationPhrases)),
 
+  inlineStyles: PropTypes.object,
   isRTL: PropTypes.bool,
 });
 
 const defaultProps = {
   disablePrev: false,
   disableNext: false,
+  navPosition: NAV_POSITION_TOP,
   navPrev: null,
   navNext: null,
   orientation: HORIZONTAL_ORIENTATION,
@@ -47,13 +53,16 @@ const defaultProps = {
 
   // internationalization
   phrases: DayPickerNavigationPhrases,
+
+  inlineStyles: null,
   isRTL: false,
 };
 
 function DayPickerNavigation({
-  css,
+  inlineStyles,
   disablePrev,
   disableNext,
+  navPosition,
   navPrev,
   navNext,
   onPrevMonthClick,
@@ -66,6 +75,8 @@ function DayPickerNavigation({
   const isHorizontal = orientation === HORIZONTAL_ORIENTATION;
   const isVertical = orientation !== HORIZONTAL_ORIENTATION;
   const isVerticalScrollable = orientation === VERTICAL_SCROLLABLE;
+  const isBottomNavPosition = navPosition === NAV_POSITION_BOTTOM;
+  const hasInlineStyles = !!inlineStyles;
 
   let navPrevIcon = navPrev;
   let navNextIcon = navNext;
@@ -73,6 +84,7 @@ function DayPickerNavigation({
   let isDefaultNavNext = false;
   let navPrevTabIndex = {};
   let navNextTabIndex = {};
+
   if (!navPrevIcon) {
     navPrevTabIndex = { tabIndex: '0' };
     isDefaultNavPrev = true;
@@ -126,6 +138,11 @@ function DayPickerNavigation({
           styles.DayPickerNavigation__verticalScrollable,
           isDefaultNav && styles.DayPickerNavigation__verticalScrollableDefault,
         ] : []),
+        ...(isBottomNavPosition ? [
+          styles.DayPickerNavigation__bottom,
+          isDefaultNav && styles.DayPickerNavigation__bottomDefault,
+        ] : []),
+        hasInlineStyles && inlineStyles,
       )}
     >
       {!isVerticalScrollable && (
@@ -140,6 +157,7 @@ function DayPickerNavigation({
               styles.DayPickerNavigation_button__horizontal,
               ...(isDefaultNavPrev ? [
                 styles.DayPickerNavigation_button__horizontalDefault,
+                isBottomNavPosition && styles.DayPickerNavigation_bottomButton__horizontalDefault,
                 !isRTL && styles.DayPickerNavigation_leftButton__horizontalDefault,
                 isRTL && styles.DayPickerNavigation_rightButton__horizontalDefault,
               ] : []),
@@ -178,6 +196,7 @@ function DayPickerNavigation({
             styles.DayPickerNavigation_button__horizontal,
             ...(isDefaultNavNext ? [
               styles.DayPickerNavigation_button__horizontalDefault,
+              isBottomNavPosition && styles.DayPickerNavigation_bottomButton__horizontalDefault,
               isRTL && styles.DayPickerNavigation_leftButton__horizontalDefault,
               !isRTL && styles.DayPickerNavigation_rightButton__horizontalDefault,
             ] : []),
@@ -238,6 +257,15 @@ export default withStyles(({ reactDates: { color, zIndex } }) => ({
     position: 'relative',
   },
 
+  DayPickerNavigation__bottom: {
+    height: 'auto',
+  },
+
+  DayPickerNavigation__bottomDefault: {
+    display: 'flex',
+    justifyContent: 'space-between',
+  },
+
   DayPickerNavigation_button: {
     cursor: 'pointer',
     userSelect: 'none',
@@ -290,6 +318,14 @@ export default withStyles(({ reactDates: { color, zIndex } }) => ({
     lineHeight: 0.78,
     borderRadius: 3,
     padding: '6px 9px',
+  },
+
+  DayPickerNavigation_bottomButton__horizontalDefault: {
+    position: 'static',
+    marginLeft: 22,
+    marginRight: 22,
+    marginBottom: 30,
+    marginTop: -10,
   },
 
   DayPickerNavigation_leftButton__horizontalDefault: {